requirements.txt
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
rich>=13.7.0
typer>=0.9.0
